debug 1

+++ include/kos/kernel.h
@@ -20,7 +20,7 @@
 	(void) (&_x == &_y);		\
 	_x > _y ? _x : _y; } )
 
-#define offsetof( TYPE, MEMBER ) ( ( size_t ) & ( (TYPE *)0 )->MEMBER )
+#define offsetof( TYPE, MEMBER ) ( (unsigned int)&( (TYPE *)0 )->MEMBER )
 
 #define container_of( ptr, type, member ) ( { \
     const typeof( ((type *)0)->member ) *__mptr = (ptr); \
